Remove formatter, fix cypher
