+++ spec/spec_helper.rb
@@ -4,10 +4,10 @@ SimpleCov.start
 require_relative 'black_fill_spec'
 require_relative 'red_fill_spec'
 require_relative 'cypher_spec'
-require_relative 'formatter_spec'
 require_relative 'bletchley_park_computer_spec'
 
 
 RSpec.configure do |config|
   config.formatter = :documentation
+  # config.profile_examples = true
 end
